travel-time-collector/requirements.txt
# 교통 소요시간 수집 서비스 (로컬/EKS)
requests>=2.31.0
python-dotenv>=1.0.0
